帖子上传接口更新,遇到一个bug,base64文本过大,数据库要换字段,明天弄

+++ 后端代码/WebCrazy2/src/main/java/app/service/postservice/PostService.java
@@ -18,12 +18,12 @@ import java.util.Map;
 public interface PostService {
 
     /**
-     * 获得帖子，按时间排序，从start开始，每次拿count个
+     * 获得帖子，按时间排序，从pageIndex开始，每次拿count个
      * @param start
      * @param count
      * @return
      */
-    List<PostInfo> getPostInfo(Long start, Long count);
+    List<PostInfo> getPostInfo(Integer start, Integer count);
 
     /**
      * 获得帖子的图片
